move download classes to output model module
