+++ antarest/study/adapters.py
@@ -15,9 +15,10 @@ from typing import Any
 import pandas as pd
 from typing_extensions import override
 
+from antarest.output.model.download import MatrixIndex
 from antarest.output.service import OutputService
 from antarest.output.storage.output_storage import OutputDetails, OutputMetadata
-from antarest.study.model import MatrixFrequency, MatrixIndex
+from antarest.study.model import MatrixFrequency
 from antarest.study.service import IOutputsAccess
 from antarest.study.storage.rawstudy.model.filesystem.inode import OriginalFile
