+++ tests/output/storage/v2/test_v2_output_storage.py
@@ -26,10 +26,11 @@ from antarest.launcher.model import LogType
 from antarest.lfs.dir_lfs import DirLargeFileStorage
 from antarest.lfs.lfs import ILargeFileStorage
 from antarest.output.filestudy.utils import MCAllAreasQueryFile
+from antarest.output.model.download import MatrixIndex
 from antarest.output.storage.v2.repository import OutputV2Repository
 from antarest.output.storage.v2.storage import V2OutputStorage
 from antarest.output.storage.v2.variables_storage import parquet_output_dir
-from antarest.study.model import MatrixFrequency, MatrixIndex, Study
+from antarest.study.model import MatrixFrequency, Study
 from antarest.study.repository import StudyMetadataRepository
 from tests.test_helpers.dates import utc_to_local
