+++ tests/study/service/test_service.py
@@ -51,6 +51,7 @@ from antarest.login.model import Group, GroupDTO, Role, User
 from antarest.login.service import LoginService
 from antarest.login.utils import current_user_context
 from antarest.matrixstore.service import MatrixService
+from antarest.output.model.download import MatrixIndex
 from antarest.output.storage.output_storage import OutputMetadata
 from antarest.study.dao.file.file_study_factory_dao import FileStudyDaoFactory
 from antarest.study.directory_service import DirectoryService
@@ -59,7 +60,6 @@ from antarest.study.model import (
     STUDY_VERSION_7_2,
     STUDY_VERSION_8,
     MatrixFrequency,
-    MatrixIndex,
     OwnerInfo,
     RawStudy,
     StorageMode,
